Add files via upload
blackjackicon.ico = an icon for the window (it is used in ClientGUI.py

ClientGUI.py = I am currently stuck on having user inputted names in a list show up on the GUI. There are many design changes as well.

Plan.txt = Just updated the plan according to what we finished.

Players.java = Used in ServerMultithread.java, in which Players is an object that contains a String name and a boolean bust. This will help keep track of who is in and out of the game, though it might be better to make Players a String, and keep track of the position in the ArrayList.

ServerMultithread.java = Just changed a few things for the Player name stuff.